:art: #132 - 프로젝트 파일의 디렉터리 수정 및 개선
:zap: #132 - CoreData 를 다루는 함수를 구현한 class 를 싱글톤으로 생성
