+++ Kloudy/Kloudy/Protocol/Protocol+.swift
