--- Kloudy/Kloudy/View/AddWeather/AddWeatherCellView/AddWeatherCellView.swift
@@ -1,8 +1,0 @@
-//
-//  AddWeatherCellView.swift
-//  Kloudy
-//
-//  Created by 이영준 on 2022/10/17.
-//
-
-import Foundation
