--- Kloudy/Kloudy/View/AddWeather/AddWeatherCellView/AddWeatherCellViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  AddWeatherCellViewModel.swift
-//  Kloudy
-//
-//  Created by 이영준 on 2022/10/17.
-//
-
-import Foundation
