--- Kloudy/Kloudy/View/AddWeather/AddWeatherPageView/AddWeatherPageView.swift
@@ -1,8 +1,0 @@
-//
-//  AddWeatherPageView.swift
-//  Kloudy
-//
-//  Created by 이영준 on 2022/10/17.
-//
-
-import Foundation
