--- Kloudy/Kloudy/View/AddWeather/AddWeatherPageView/AddWeatherPageViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  AddWeatherPageViewModel.swift
-//  Kloudy
-//
-//  Created by 이영준 on 2022/10/17.
-//
-
-import Foundation
