--- Kloudy/Kloudy/View/AddWeather/SearchWeatherCellView/SearchWeatherView.swift
@@ -1,8 +1,0 @@
-//
-//  SearchWeatherView.swift
-//  Kloudy
-//
-//  Created by 이영준 on 2022/10/17.
-//
-
-import Foundation
