--- Kloudy/Kloudy/View/AddWeather/SearchWeatherCellView/SearchWeatherViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  SearchWeatherViewModel.swift
-//  Kloudy
-//
-//  Created by 이영준 on 2022/10/17.
-//
-
-import Foundation
