--- Kloudy/Kloudy/View/AddWeather/SearchWeatherGridView/SearchWeatherGridView.swift
@@ -1,8 +1,0 @@
-//
-//  SearchWeatherGridView.swift
-//  Kloudy
-//
-//  Created by 이영준 on 2022/10/17.
-//
-
-import Foundation
