--- Kloudy/Kloudy/View/AddWeather/SearchWeatherGridView/SearchWeatherGridViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  SearchWeatherGridViewModel.swift
-//  Kloudy
-//
-//  Created by 이영준 on 2022/10/17.
-//
-
-import Foundation
