--- Kloudy/Kloudy/View/AddWeather/SearchWeatherListView/SearchWeatherListView.swift
@@ -1,8 +1,0 @@
-//
-//  SearchWeatherListView.swift
-//  Kloudy
-//
-//  Created by 이영준 on 2022/10/17.
-//
-
-import Foundation
