--- Kloudy/Kloudy/View/AddWeather/SearchWeatherListView/SearchWeatherListViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  SearchWeatherListViewModel.swift
-//  Kloudy
-//
-//  Created by 이영준 on 2022/10/17.
-//
-
-import Foundation
