--- Kloudy/Kloudy/View/AddWeather/SelectWeatherCellView/SelectWeatherCellView.swift
@@ -1,16 +1,0 @@
-//
-//  SelectWeatherCellView.swift
-//  Kloudy
-//
-//  Created by 이영준 on 2022/10/17.
-//
-
-import UIKit
-import SnapKit
-
-class SelectWeatherCellView: UIViewController {
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        self.view.backgroundColor = .white
-    }
-}
