--- Kloudy/Kloudy/View/AddWeather/SelectWeatherCellView/SelectWeatherCellViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  SelectWeatherCellViewModel.swift
-//  Kloudy
-//
-//  Created by 이영준 on 2022/10/17.
-//
-
-import Foundation
