--- Kloudy/Kloudy/View/CheckLocation/LocationSelectionEditView/LocationSelectionEditViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  LocationSelectionEditViewModel.swift
-//  Kloudy
-//
-//  Created by 이영준 on 2022/10/17.
-//
-
-import Foundation
