+++ Kloudy/Kloudy/View/CheckLocation/LocationSelectionView/SearchLocationView.swift
@@ -1,0 +1,8 @@
+//
+//  SearchLocationView.swift
+//  Kloudy
+//
+//  Created by 이영준 on 2022/11/07.
+//
+
+import Foundation
