--- Kloudy/Kloudy/View/CheckWeather/CheckCityWeatherView/CheckCityView.swift
@@ -1,8 +1,0 @@
-//
-//  CheckCityView.swift
-//  Kloudy
-//
-//  Created by 이영준 on 2022/10/17.
-//
-
-import Foundation
