--- Kloudy/Kloudy/View/CheckWeather/CheckCityWeatherView/CheckCityViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  CheckCityViewModel.swift
-//  Kloudy
-//
-//  Created by 이영준 on 2022/10/17.
-//
-
-import Foundation
