--- Kloudy/Kloudy/View/CheckWeather/CheckWeatherCollectionView/CheckWeatherCollectionView.swift
@@ -1,10 +1,0 @@
-//
-//  CheckWeatherCollectionView.swift
-//  Kloudy
-//
-//  Created by 이영준 on 2022/10/17.
-//
-
-
-import UIKit
-
