--- Kloudy/Kloudy/View/CheckWeather/CheckWeatherCollectionView/CheckWeatherCollectionViewModel.swift
@@ -1,20 +1,0 @@
-//
-//  CheckWeatherCollectionViewModel.swift
-//  Kloudy
-//
-//  Created by 이영준 on 2022/10/17.
-//
-//
-//import Foundation
-//
-//struct CheckWeatherCollectionViewModel {
-//    var todayIndexArray: [cellData] = [
-//        cellData(indexName: "01.circle"),
-//        cellData(indexName: "02.circle"),
-//        cellData(indexName: "03.circle"),
-//        cellData(indexName: "04.circle"),
-//        cellData(indexName: "05.circle"),
-//        cellData(indexName: "06.circle"),
-//        cellData(indexName: "07.circle")
-//    ]
-//}
