--- Kloudy/Kloudy/View/CheckWeather/CheckWeatherEditNavigationView/CheckWeatherEditNavigationViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  CheckWeatherEditNavigationViewModel.swift
-//  Kloudy
-//
-//  Created by 이영준 on 2022/10/17.
-//
-
-import Foundation
