--- Kloudy/Kloudy/View/CheckWeather/CheckWeatherPageView/CheckWeatherPageViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  CheckWeatherPageViewModel.swift
-//  Kloudy
-//
-//  Created by 이영준 on 2022/10/17.
-//
-
-import Foundation
