--- Kloudy/Kloudy/View/WeatherCells/WeatherCellFourByFour/WeatherCellFourByFour.swift
@@ -1,8 +1,0 @@
-//
-//  WeatherCellFourByFour.swift
-//  Kloudy
-//
-//  Created by 이영준 on 2022/10/17.
-//
-
-import Foundation
