--- Kloudy/Kloudy/View/WeatherCells/WeatherCellFourByFour/WeatherCellFourByFourViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  WeatherCellFourByFourViewModel.swift
-//  Kloudy
-//
-//  Created by 이영준 on 2022/10/17.
-//
-
-import Foundation
