--- Kloudy/Kloudy/View/WeatherCells/WeatherCellFourByOne/WeatherCellFourByOne.swift
@@ -1,8 +1,0 @@
-//
-//  WeatherCellFourByOne.swift
-//  Kloudy
-//
-//  Created by 이영준 on 2022/10/17.
-//
-
-import Foundation
