--- Kloudy/Kloudy/View/WeatherCells/WeatherCellFourByOne/WeatherCellFourByOneViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  WeatherCellFourByOneViewModel.swift
-//  Kloudy
-//
-//  Created by 이영준 on 2022/10/17.
-//
-
-import Foundation
