--- Kloudy/Kloudy/View/WeatherCells/WeatherCellFourByTwo/WeatherCellFourByTwo.swift
@@ -1,8 +1,0 @@
-//
-//  WeatherCellFourByTwo.swift
-//  Kloudy
-//
-//  Created by 이영준 on 2022/10/17.
-//
-
-import Foundation
