--- Kloudy/Kloudy/ViewController.swift
@@ -1,18 +1,0 @@
-//
-//  ViewController.swift
-//  Kloudy
-//
-//  Created by 이영준 on 2022/10/13.
-//
-
-import UIKit
-import CoreData
-
-class ViewController: UIViewController {
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-}
